# WeatherStation_Backend/src/main/java/com/tony/wheatherstation/controller/AuthController.java
package com.tony.wheatherstation.controller;

import com.tony.wheatherstation.dto.auth.AuthResponse;
import com.tony.wheatherstation.dto.auth.LoginRequest;
import com.tony.wheatherstation.dto.auth.RefreshRequest;
import com.tony.wheatherstation.dto.auth.RegisterRequest;
import com.tony.wheatherstation.service.AuthService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

/** Endpoints públicos de autenticación (US1). */
@RestController
@RequestMapping("/auth")
@RequiredArgsConstructor
@Tag(name = "Auth", description = "Registro, login y refresh de tokens")
public class AuthController {

    private final AuthService authService;

    @Operation(summary = "Registrar un nuevo usuario (rol USUARIO)")
    @PostMapping("/register")
    public ResponseEntity<AuthResponse> register(@Valid @RequestBody RegisterRequest request) {
        return ResponseEntity.status(HttpStatus.CREATED).body(authService.register(request));
    }

    @Operation(summary = "Iniciar sesión")
    @PostMapping("/login")
    public AuthResponse login(@Valid @RequestBody LoginRequest request) {
        return authService.login(request);
    }

    @Operation(summary = "Renovar el access token (rota el refresh token)")
    @PostMapping("/refresh")
    public AuthResponse refresh(@Valid @RequestBody RefreshRequest request) {
        return authService.refresh(request);
    }
}