# WeatherStation_Backend/src/main/java/com/tony/wheatherstation/dto/common/ApiError.java
package com.tony.wheatherstation.dto.common;

import com.fasterxml.jackson.annotation.JsonInclude;

import java.time.Instant;
import java.util.Map;

/**
 * Cuerpo de error uniforme para toda la API (FR-026). No expone trazas internas
 * ni secretos. {@code fieldErrors} solo se incluye en errores de validación.
 */
@JsonInclude(JsonInclude.Include.NON_NULL)
public record ApiError(
        Instant timestamp,
        int status,
        String error,
        String code,
        String message,
        String path,
        Map<String, String> fieldErrors
) {
    public static ApiError of(int status, String error, String code, String message, String path) {
        return new ApiError(Instant.now(), status, error, code, message, path, null);
    }

    public static ApiError of(int status, String error, String code, String message, String path,
                              Map<String, String> fieldErrors) {
        return new ApiError(Instant.now(), status, error, code, message, path, fieldErrors);
    }
}