# specs/001-weatherstation-backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WeatherStation Backend API (Red de Estaciones)
  description: >
    API REST del backend central de la Red de Estaciones Meteorológicas
    Inteligentes (CLIMBOT). Único componente con acceso a la base de datos
    PostgreSQL (almacén único: sistema + lecturas meteorológicas) y a Gemini (IA).
    Las estaciones ESP32 publican sus lecturas a través de `/api/device/**` (nunca
    a un almacén externo). Autenticación JWT (Bearer) con roles ADMIN, RESPONSABLE,
    INVESTIGADOR y USUARIO para usuarios, y token de estación → JWT de dispositivo
    para los ESP32.
  version: 3.1.0
  contact:
    name: WeatherStation
servers:
  - url: http://localhost:8080
    description: Desarrollo local
  - url: https://weatherstation-backend.onrender.com
    description: Producción (Render)
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar un nuevo usuario (rol USUARIO)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Usuario creado y autenticado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Autenticación correcta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar el access token (rota el refresh token)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: Nuevos tokens emitidos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ─────────────────────── DISPOSITIVO (ESP32) ───────────────────────
  /api/device/register:
    post:
      tags: [Dispositivo]
      summary: Solicitud de alta iniciada por el dispositivo (queda PENDING)
      description: >
        Un ESP32 no aprovisionado solicita su registro. Crea una solicitud
        PENDING; no devuelve token hasta que un ADMIN apruebe. Rate limited.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EstacionRegistroRequest' }
      responses:
        '202':
          description: Solicitud registrada (pendiente de aprobación)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SolicitudResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/device/auth:
    post:
      tags: [Dispositivo]
      summary: Handshake de la estación (token permanente → JWT de dispositivo)
      description: >
        Valida que la estación exista y esté APPROVED y que el token sea válido.
        Devuelve un JWT de dispositivo de vida corta. Registra la conexión.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceAuthRequest' }
      responses:
        '200':
          description: Autenticación de estación correcta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceAuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Estación no aprobada (PENDING/REJECTED/DISABLED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/device/data:
    post:
      tags: [Dispositivo]
      summary: Publicar una lectura (el backend la valida y la persiste en PostgreSQL)
      description: >
        Requiere el JWT de dispositivo obtenido en /api/device/auth. El backend
        valida formato y rangos, persiste en PostgreSQL (lectura_actual y, según
        cadencia, el histórico lecturas) y actualiza la última conexión.
      security:
        - deviceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceDataRequest' }
      responses:
        '202':
          description: Lectura aceptada y persistida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceDataResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Estación no autorizada para publicar
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '422':
          description: Lectura sintácticamente válida pero fuera de rango
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/device/heartbeat:
    post:
      tags: [Dispositivo]
      summary: Latido de salud de la estación (firmware, hardware, RSSI; sin batería)
      description: >
        Requiere el JWT de dispositivo. Actualiza la última conexión y permite
        derivar ONLINE/OFFLINE. No recolecta telemetría de batería.
      security:
        - deviceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceHeartbeatRequest' }
      responses:
        '204': { description: Latido registrado }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/device/config:
    get:
      tags: [Dispositivo]
      summary: Configuración remota de la estación (intervalo, muestreo, zona horaria)
      security:
        - deviceAuth: []
      responses:
        '200':
          description: Configuración vigente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceConfigResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ──────────────────────── PÚBLICO (sin auth) ───────────────────────
...
  /escuelas:
    get:
      tags: [Escuelas]
      summary: Listar escuelas (autenticado)
      responses:
        '200':
          description: Lista de escuelas
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EscuelaResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Escuelas]
      summary: Crear escuela (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EscuelaRequest' }
      responses:
        '201':
          description: Escuela creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EscuelaResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /escuelas/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Escuelas]
      summary: Ver una escuela (autenticado)
      responses:
        '200':
          description: Escuela
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EscuelaResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Escuelas]
      summary: Actualizar una escuela (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EscuelaRequest' }
      responses:
        '200':
          description: Escuela actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EscuelaResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Escuelas]
      summary: Eliminar una escuela (ADMIN)
      responses:
        '204': { description: Eliminada }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Escuela con estaciones asociadas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }

  # ─────────────────────── ESTACIONES (gobernanza) ───────────────────
...
  /alertas:
    get:
      tags: [Alertas]
      summary: Listar alertas (cualquier rol)
      parameters:
        - { in: query, name: estacionId, schema: { type: string, format: uuid } }
        - { in: query, name: tipo, schema: { $ref: '#/components/schemas/TipoAlerta' } }
        - { in: query, name: estado, schema: { $ref: '#/components/schemas/EstadoAlerta' } }
      responses:
        '200':
          description: Lista de alertas
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AlertaResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /alertas/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Alertas]
      summary: Detalle de una alerta (cualquier rol)
      responses:
        '200':
          description: Alerta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlertaResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de usuario (claim `rol`).
    deviceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de dispositivo emitido por /api/device/auth (claim `type=DEVICE`).
...
  responses:
    BadRequest:
      description: Datos de entrada inválidos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: No autenticado o credenciales inválidas
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Rol/recurso no autorizado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Conflicto (p. ej. email/clave ya existe)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    TooManyRequests:
      description: Límite de peticiones excedido (IA o dispositivo)
      headers:
        Retry-After:
          schema: { type: integer }
          description: Segundos hasta poder reintentar
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ServiceUnavailable:
      description: Servicio externo (Gemini) o base de datos no disponible
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    Rol:
      type: string
      enum: [ADMIN, RESPONSABLE, INVESTIGADOR, USUARIO]
...
    EstadoSolicitud:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    TipoAlerta:
      type: string
      enum: [LLUVIA, VIENTO_FUERTE, CALOR_EXTREMO, ESTACION_DESCONECTADA, SENSOR_SIN_RESPUESTA]
    Severidad:
      type: string
      enum: [BAJA, MEDIA, ALTA]
    EstadoAlerta:
      type: string
      enum: [ACTIVA, RESUELTA]
    RegisterRequest:
      type: object
      required: [nombre, email, password]
      properties:
        nombre: { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 72, format: password }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresIn: { type: integer, description: Segundos de vida del access token }
        usuario: { $ref: '#/components/schemas/UsuarioResponse' }
...
    UsuarioResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        email: { type: string, format: email }
        rol: { $ref: '#/components/schemas/Rol' }
        escuelaId: { oneOf: [ { type: string, format: uuid }, { type: 'null' } ] }
        escuelaNombre: { oneOf: [ { type: string }, { type: 'null' } ] }
        activo: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    EscuelaRequest:
      type: object
      required: [nombre]
      properties:
        nombre: { type: string, minLength: 2, maxLength: 150 }
        clave: { type: string, description: CCT u otra clave oficial }
        municipio: { type: string }
        ubicacion: { type: string }
        latitud: { type: number, format: double }
        longitud: { type: number, format: double }
        contactoEmail: { type: string, format: email }
    EscuelaResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        clave: { type: string }
        municipio: { type: string }
        ubicacion: { type: string }
        latitud: { type: number, format: double }
        longitud: { type: number, format: double }
        contactoEmail: { type: string, format: email }
        totalEstaciones: { type: integer }
        createdAt: { type: string, format: date-time }
    EstacionRegistroRequest:
      type: object
      required: [nombre, escuelaId]
      properties:
        nombre: { type: string, minLength: 2, maxLength: 120 }
        escuelaId: { type: string, format: uuid }
        ubicacion: { type: string }
        municipio: { type: string }
        latitud: { type: number, format: double }
        longitud: { type: number, format: double }
        firmware: { type: string, maxLength: 40 }
...
    SolicitudResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        uuidPropuesto: { type: string, format: uuid }
        nombre: { type: string }
        escuelaId: { type: string, format: uuid }
        escuelaNombre: { type: string }
        ubicacion: { type: string }
        municipio: { type: string }
        firmware: { type: string }
        estado: { $ref: '#/components/schemas/EstadoSolicitud' }
        motivoRechazo: { oneOf: [ { type: string }, { type: 'null' } ] }
        estacionId: { oneOf: [ { type: string, format: uuid }, { type: 'null' } ] }
        createdAt: { type: string, format: date-time }
        resueltaEn: { oneOf: [ { type: string, format: date-time }, { type: 'null' } ] }
    DeviceAuthRequest:
      type: object
      required: [uuid, token]
      properties:
        uuid: { type: string, format: uuid }
        token: { type: string, description: Token permanente de la estación }
        firmware: { type: string, maxLength: 40 }
    DeviceAuthResponse:
      type: object
      properties:
        deviceToken: { type: string, description: JWT de dispositivo de vida corta }
        tokenType: { type: string, example: Bearer }
        expiresIn: { type: integer, description: Segundos de vida del JWT de dispositivo }
    DeviceDataRequest:
      type: object
      required: [timestamp, temperatura, humedad, presion, vientoKmh, lluviaMm]
      properties:
        timestamp: { type: string, format: date-time }
        temperatura: { type: number, format: double, example: 24.5 }
        humedad: { type: number, format: double, minimum: 0, maximum: 100, example: 60 }
        presion: { type: number, format: double, minimum: 0.1, example: 1013.2 }
        vientoKmh: { type: number, format: double, minimum: 0, example: 12.3 }
        vientoDir: { type: string, example: NE }
        vientoGrados: { type: number, format: double, minimum: 0, maximum: 360, example: 45.0 }
        lluviaMm: { type: number, format: double, minimum: 0, example: 0.0 }
    DeviceDataResponse:
      type: object
      properties:
        estado: { type: string, example: ACEPTADA }
        recibidoEn: { type: string, format: date-time }
        historialActualizado: { type: boolean }
    DeviceHeartbeatRequest:
      type: object
      description: Latido de salud. SIN telemetría de batería.
      properties:
        firmware: { type: string, maxLength: 40 }
        hardware: { type: string, maxLength: 60 }
        rssi: { type: integer, description: WiFi.RSSI() en dBm }
        uptimeSeg: { type: integer, format: int64 }
    DeviceConfigResponse:
      type: object
      properties:
        intervaloEnvioSeg: { type: integer, example: 60 }
        muestreoSeg: { type: integer, example: 5 }
        zonaHoraria: { type: string, example: America/Mexico_City }
        servidorAhora: { type: string, format: date-time }
...
    AlertaResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        estacionId: { type: string, format: uuid }
        tipo: { $ref: '#/components/schemas/TipoAlerta' }
        severidad: { $ref: '#/components/schemas/Severidad' }
        estado: { $ref: '#/components/schemas/EstadoAlerta' }
        mensaje: { type: string }
        valorDisparo: { type: number, format: double }
        variableDisparo: { type: string }
        detectadaEn: { type: string, format: date-time }
        resueltaEn:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
    ApiError:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        code: { type: string, example: ESTACION_NO_APROBADA }
        message: { type: string }
        path: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }